Move diff types into own files
