+++ coding_systems/icd10/known_diffs/__init__.py
@@ -1,2 +1,25 @@
-from .known_diffs import *
+# A record of known differences between:
+# - 2016 CLAML from WHO website and the 2016 CLAML generated by scraping the NHS classbrowser
+# - the NHS specific logic for things like place of occurrence modifiers
+# - the 2016 scraped and 2019 CLAML from WHO website
 
+# ruff: noqa: F401
+
+from .claml_vs_scraped import (
+    is_2016_claml_only,
+    is_2016_description_difference,
+    is_2016_scraped_only,
+    should_include_2016_claml_only,
+    should_include_2016_scraped_only,
+    should_use_scraped_for_2016,
+)
+from .combined_2016_vs_2019 import (
+    COMBINED_2016_VS_2019_DIFFERENCES,
+    clinically_different_codes,
+    codes_with_different_descriptions,
+    get_2016_2019_description_difference,
+)
+from .differences import ReleaseTermDifference, RubricChange, TermDifference
+from .moved import moved_codes
+from .rubrics import KNOWN_2016_RUBRIC_CHANGES
+from .who_2016_overrides import expand_who_2016_place_of_occurrence
