--- coding_systems/icd10/known_diffs/known_diffs.py
@@ -1,1496 +1,0 @@
-# A record of known differences between:
-# - 2016 CLAML from WHO website and the 2016 CLAML generated by scraping the NHS classbrowser
-# - the NHS specific logic for things like place of occurrence modifiers
-# - the 2016 scraped and 2019 CLAML from WHO website
-
-import re
-from dataclasses import dataclass, field
-
-from coding_systems.icd10.claml_parser import ICD10Code, ModifierDigit
-
-
-@dataclass(frozen=True, slots=True)
-class CodeDifference:
-    include_in_release: bool
-
-
-@dataclass(frozen=True, slots=True)
-class TermDifference:
-    claml: str
-    scraped: str
-    use: str
-
-    def __post_init__(self) -> None:
-        if self.use not in ("claml", "scraped"):
-            raise ValueError(f"use must be 'claml' or 'scraped', got {self.use!r}")
-
-
-@dataclass(frozen=True, slots=True)
-class KnownDifferences:
-    claml_only: dict[str, CodeDifference]
-    scraped_only: dict[str, CodeDifference]
-    term_differences: dict[str, TermDifference]
-
-
-@dataclass(frozen=True, slots=True)
-class ReleaseTermDifference:
-    combined_2016: str
-    who_2019: str
-    clinically_equivalent: bool
-
-
-WHO_2016_EXPECTED_OVERRIDES = frozenset(
-    # Taken from https://classbrowser.nhs.uk/ICD-10-5TH-Edition/NCCS.pdf#page=218
-    #
-    # A fourth character must be assigned with codes from categories W00-Y34 to identify
-    # where the injury, poisoning or adverse effect took place. The fourth characters can
-    # be found in the ‘Place of occurrence code’ section at the beginning of the chapter.
-    # The exceptions are codes in categories Y06.- Neglect and abandonment and
-    # Y07.- Other maltreatment.
-    #
-    # However it then goes on to say in
-    #  https://classbrowser.nhs.uk/ICD-10-5TH-Edition/NCCS.pdf#page=221
-    #
-    # The fourth character codes printed at categories W26, X34, and X59 in the
-    # ICD-10 Tabular List must not be used and must be crossed through in the ICD-10 5th
-    # Edition books.
-    # The ‘Place of occurrence codes’ must be used for fourth character code assignment with
-    # categories W26.-, X34.- and X59.-.
-    # The WHO Update Revision Committee previously agreed that ‘Place of occurrence codes’
-    # should be separated from the three character Chapter XX code. Following this agreement
-    # WHO introduced new fourth character subcategory codes at W26.-, X34.- and X59.- . This
-    # re-designation and re-use of the fourth character by the WHO is incompatible with current
-    # UK practice when using categories W26.-, X34.- and X59.-, therefore the implementation
-    # of these fourth character subcategories has been deferred.
-    #
-    # This means when applying the place of occurrence modifiers we would expect the following
-    # actual 4 character children to be overridden by the place of occurrence modifiers. We
-    # want to confirm that these are the only ones, and that no other unexpected codes are
-    # generated by the place of occurrence modifiers.
-    {"W260", "W268", "W269", "X340", "X341", "X348", "X349", "X590", "X599"}
-)
-
-
-def expand_who_2016_place_of_occurrence(
-    records: dict[str, ICD10Code],
-    place_modifiers: list[ModifierDigit],
-) -> dict[str, ICD10Code]:
-    """
-    Expand the given records with place of occurrence modifiers for the 2016 WHO release,
-    and check that the expected overrides are present and no unexpected codes are generated.
-    """
-
-    place_records = {}
-    for base_code, base_record in records.items():
-        # We only apply to 3 character codes in the specified range
-        if not re.match(r"^[A-Z][0-9]{2}$", base_code):
-            continue
-
-        # We only apply to codes in the range W00-Y34
-        if not ("W00" <= base_code <= "Y34"):
-            continue
-
-        # Apart from the exceptions Y06 and Y07
-        if base_code in {"Y06", "Y07"}:
-            continue
-
-        for modifier in place_modifiers:
-            new_code = base_code + modifier.digit_code
-            place_records[new_code] = ICD10Code(
-                code=new_code,
-                parent=base_code,
-                description=base_record.description,
-                term_modifier=modifier.description,
-                modifier_position=4,
-            )
-
-    overlap = set(place_records) & set(records)
-    unexpected_overlap = overlap - WHO_2016_EXPECTED_OVERRIDES
-    assert not unexpected_overlap, (
-        "Place-of-occurrence generated codes overlap with unexpected existing codes: "
-        f"{sorted(unexpected_overlap)}"
-    )
-    unexpected_missing = WHO_2016_EXPECTED_OVERRIDES - set(place_records)
-    assert not unexpected_missing, (
-        "Place-of-occurrence expected codes missing from generated codes: "
-        f"{sorted(unexpected_missing)}"
-    )
-    return place_records
-
-
-# Known differences between the 2016 WHO ICD10 claml and the scraped ICD10
-# data from the NHS class browser
-CLAML_VS_SCRAPED_DIFFERENCES = KnownDifferences(
-    claml_only={
-        # The CLAML says you can apply the 5th char modifier to M14.*, but
-        # doesn't say you can in any online version. It also doesn't appear
-        # in the backend data. But no harm in including, and it is technically
-        # in the XML, so we keep it. The 5th char modifiers are 0-9 and apply
-        # to M140-M146 and M148-M149, but not M147 which doesn't exist
-        **{
-            f"M{i}": CodeDifference(include_in_release=True)
-            for i in list(range(1400, 1470)) + list(range(1480, 1490))
-        },
-        # W00-Y34 (except Y06 and Y07) are all 3 char codes in the CLAML,
-        # which the NHS allows a 4th character "place of occurrence" modifier.
-        # The online browser metnions this, but not in a structure way so
-        # the scraper doesn't pick it up. We therefore expect all codes
-        # in that range to be in the claml but not the scraped data.
-        # W000-W999
-        **{
-            f"W{number:03d}": CodeDifference(include_in_release=True)
-            for number in range(0, 1000)
-        },
-        # X000-X999
-        **{
-            f"X{number:03d}": CodeDifference(include_in_release=True)
-            for number in range(0, 1000)
-        },
-        # Y000-Y349 (except Y06 and Y07)
-        **{
-            f"Y{number:03d}": CodeDifference(include_in_release=True)
-            for number in list(range(0, 600)) + list(range(800, 3500))
-        },
-    },
-    scraped_only={
-        # Currently no codes that only appear in the scraped data but keeping
-        # the property here in case we want to add some in future
-    },
-    term_differences={
-        # The expected differences in the terms between the CLAML and scraped
-        # data, and which one we want to use in the release. The scraped term
-        # is usually the one we prefer as we're trying to match the NHS classbrowser
-        # as closely as possible, but in some cases the CLAML term is better.
-        "J10": TermDifference(
-            claml="influenza due to identified seasonal influenza virus",
-            scraped="Influenza due to identified seasonal influenza virus",
-            use="scraped",
-        ),
-        "P710": TermDifference(
-            claml="Cow milk hypocalcaemia in newborn",
-            scraped="Cow's milk hypocalcaemia in newborn",
-            use="scraped",
-        ),
-        "U06": TermDifference(
-            claml="Zika virus disease",
-            scraped="Emergency use of U06",
-            use="claml",
-            # U06 isn't a code you can actually use (as it has 4 char children) so
-            # the NHS haven't updated it. So the claml version is more accurate
-            # now that U609 is Zika virus instead of emergency use
-        ),
-        "U069": TermDifference(
-            claml="Zika virus disease, unspecified",
-            scraped="Zika virus disease",
-            use="scraped",
-        ),
-        "U070": TermDifference(
-            claml="Emergency use of U07.0",
-            scraped="Vaping-related disorder",
-            use="scraped",
-        ),
-        "U071": TermDifference(
-            claml="Emergency use of U07.1",
-            scraped="COVID-19, virus identified",
-            use="scraped",
-        ),
-        "U072": TermDifference(
-            claml="Emergency use of U07.2",
-            scraped="COVID-19, virus not identified",
-            use="scraped",
-        ),
-        "U073": TermDifference(
-            claml="Emergency use of U07.3",
-            scraped="Personal history of COVID-19",
-            use="scraped",
-        ),
-        "U074": TermDifference(
-            claml="Emergency use of U07.4",
-            scraped="Post COVID-19 condition",
-            use="scraped",
-        ),
-        "U075": TermDifference(
-            claml="Emergency use of U07.5",
-            scraped="Multisystem inflammatory syndrome associated with COVID-19",
-            use="scraped",
-        ),
-        "U076": TermDifference(
-            claml="Emergency use of U07.6",
-            scraped="Need for immunization against COVID-19",
-            use="scraped",
-        ),
-        "U077": TermDifference(
-            claml="Emergency use of U07.7",
-            scraped="COVID-19 vaccines causing adverse effects in therapeutic use",
-            use="scraped",
-        ),
-    },
-)
-
-
-# Known differences between the combined 2016 data (base WHO claml + scraped data)
-# and the 2019 WHO claml. We no longer care about codes in one but not the other as
-# our model supports that. The model also supports different descriptions for the
-# same code - but we list them all here so we can (a) check that no unexpected
-# differences exist - particularly for future releases, and (b) classify whether the
-# differences are clinically equivalent or not, which is important for the user
-# interface to show a warning to users
-COMBINED_2016_VS_2019_DIFFERENCES = {
-    # Norwark agent is another name for Norovirus, so clinically equivalent
-    "A081": ReleaseTermDifference(
-        combined_2016="Acute gastroenteropathy due to Norwalk agent",
-        who_2019="Acute gastroenteropathy due to Norovirus",
-        clinically_equivalent=True,
-    ),
-    # Enterococcus used to be considered a group D streptococcus, but is now
-    # considered a separate genus. All other instances of "enterococcus" in
-    # all ICD10 editions appear as "streptococcus group D and enterococcus"
-    # so I think we can count these as clinically equivalent
-    "A402": ReleaseTermDifference(
-        combined_2016="Sepsis due to streptococcus, group D",
-        who_2019="Sepsis due to streptococcus, group D and enterococcus",
-        clinically_equivalent=True,
-    ),
-    # Confirmed with clinician that these are equivalent:
-    # https://bennettoxford.slack.com/archives/C03HVM72273/p1782299456089129
-    "B170": ReleaseTermDifference(
-        combined_2016="Acute delta-(super)infection of hepatitis B carrier",
-        who_2019="Acute delta-(super)infection in chronic hepatitis B",
-        clinically_equivalent=True,
-    ),
-    # The 2019 term has an inclusion rubric of "Tropical spastic paraplegia"
-    # so these are the same
-    "G041": ReleaseTermDifference(
-        combined_2016="Tropical spastic paraplegia",
-        who_2019="Human T-cell lymphotropic virus associated myelopathy",
-        clinically_equivalent=True,
-    ),
-    # Clinician confirmed that these could be different
-    # https://bennettoxford.slack.com/archives/C03HVM72273/p1782299456089129
-    "I620": ReleaseTermDifference(
-        combined_2016="Subdural haemorrhage (acute)(nontraumatic)",
-        who_2019="Nontraumatic subdural haemorrhage",
-        clinically_equivalent=False,
-    ),
-    # Small typographical change only, so clinically equivalent
-    "P710": ReleaseTermDifference(
-        combined_2016="Cow's milk hypocalcaemia in newborn",
-        who_2019="Cow milk hypocalcaemia in newborn",
-        clinically_equivalent=True,
-    ),
-    # Based on the following from the NHS guidance it looks like the 2016 term was
-    # incorrectly missing the "congenital" bit. Or at least the "congenital" was
-    # inferred from the ancestor codes of Q39.4
-    #
-    # The ICD-10 Alphabetical Index assumes that an oesophageal web is a congenital condition
-    # and classifies this to Q39.4 Oesophageal web. However, an oesophageal web can be
-    # either congenital or acquired, with acquired being more common.
-    # The following must be applied when coding oesophageal web:
-    #   • A documented diagnosis of congenital oesophageal web must be classified to Q39.4
-    #     Oesophageal web.
-    #   • A documented diagnosis of acquired oesophageal web must be classified to
-    #     K22.2 Oesophageal obstruction.
-    #   • An unspecified oesophageal web (i.e. not documented as congenital or acquired)
-    #     must be classified to K22.2 Oesophageal obstruction.
-    "Q394": ReleaseTermDifference(
-        combined_2016="Oesophageal web",
-        who_2019="Congenital oesophageal web",
-        clinically_equivalent=True,
-    ),
-    # Clinician confirmed that these could be different
-    # https://bennettoxford.slack.com/archives/C03HVM72273/p1782299456089129
-    "R17": ReleaseTermDifference(
-        combined_2016="Unspecified jaundice",
-        who_2019="Hyperbilirubinaemia, with or without jaundice, not elsewhere classified",
-        clinically_equivalent=False,
-    ),
-    # The 2019 version has an inclusion rubric which specifies "due to self neglect"
-    # so these are clinically equivalent
-    "R636": ReleaseTermDifference(
-        combined_2016="Insufficient intake of food and water due to self neglect",
-        who_2019="Insufficient intake of food and water",
-        clinically_equivalent=True,
-    ),
-    # Minor change, but clinically equivalent
-    "T602": ReleaseTermDifference(
-        combined_2016="Toxic effect: Other insecticides",
-        who_2019="Toxic effect: Other and unspecified insecticides",
-        clinically_equivalent=True,
-    ),
-    # Extra "other" but otherwise equivalent
-    "T758": ReleaseTermDifference(
-        combined_2016="Other specified effects of external causes",
-        who_2019="Other specified effects of other external causes",
-        clinically_equivalent=True,
-    ),
-    # W20 and W22: just the addition of "(s)" to "object" in the term, so clinically equivalent
-    "W20": ReleaseTermDifference(
-        combined_2016="Struck by thrown, projected or falling object",
-        who_2019="Struck by thrown, projected or falling object(s)",
-        clinically_equivalent=True,
-    ),
-    "W200": ReleaseTermDifference(
-        combined_2016="Struck by thrown, projected or falling object (Home)",
-        who_2019="Struck by thrown, projected or falling object(s) (Home)",
-        clinically_equivalent=True,
-    ),
-    "W201": ReleaseTermDifference(
-        combined_2016="Struck by thrown, projected or falling object (Residential institution)",
-        who_2019="Struck by thrown, projected or falling object(s) (Residential institution)",
-        clinically_equivalent=True,
-    ),
-    "W202": ReleaseTermDifference(
-        combined_2016="Struck by thrown, projected or falling object (School, other institution and public administrative area)",
-        who_2019="Struck by thrown, projected or falling object(s) (School, other institution and public administrative area)",
-        clinically_equivalent=True,
-    ),
-    "W203": ReleaseTermDifference(
-        combined_2016="Struck by thrown, projected or falling object (Sports and athletics area)",
-        who_2019="Struck by thrown, projected or falling object(s) (Sports and athletics area)",
-        clinically_equivalent=True,
-    ),
-    "W204": ReleaseTermDifference(
-        combined_2016="Struck by thrown, projected or falling object (Street and highway)",
-        who_2019="Struck by thrown, projected or falling object(s) (Street and highway)",
-        clinically_equivalent=True,
-    ),
-    "W205": ReleaseTermDifference(
-        combined_2016="Struck by thrown, projected or falling object (Trade and service area)",
-        who_2019="Struck by thrown, projected or falling object(s) (Trade and service area)",
-        clinically_equivalent=True,
-    ),
-    "W206": ReleaseTermDifference(
-        combined_2016="Struck by thrown, projected or falling object (Industrial and construction area)",
-        who_2019="Struck by thrown, projected or falling object(s) (Industrial and construction area)",
-        clinically_equivalent=True,
-    ),
-    "W207": ReleaseTermDifference(
-        combined_2016="Struck by thrown, projected or falling object (Farm)",
-        who_2019="Struck by thrown, projected or falling object(s) (Farm)",
-        clinically_equivalent=True,
-    ),
-    "W208": ReleaseTermDifference(
-        combined_2016="Struck by thrown, projected or falling object (Other specified places)",
-        who_2019="Struck by thrown, projected or falling object(s) (Other specified places)",
-        clinically_equivalent=True,
-    ),
-    "W209": ReleaseTermDifference(
-        combined_2016="Struck by thrown, projected or falling object (Unspecified place)",
-        who_2019="Struck by thrown, projected or falling object(s) (Unspecified place)",
-        clinically_equivalent=True,
-    ),
-    "W22": ReleaseTermDifference(
-        combined_2016="Striking against or struck by other objects",
-        who_2019="Striking against or struck by other object(s)",
-        clinically_equivalent=True,
-    ),
-    "W220": ReleaseTermDifference(
-        combined_2016="Striking against or struck by other objects (Home)",
-        who_2019="Striking against or struck by other object(s) (Home)",
-        clinically_equivalent=True,
-    ),
-    "W221": ReleaseTermDifference(
-        combined_2016="Striking against or struck by other objects (Residential institution)",
-        who_2019="Striking against or struck by other object(s) (Residential institution)",
-        clinically_equivalent=True,
-    ),
-    "W222": ReleaseTermDifference(
-        combined_2016="Striking against or struck by other objects (School, other institution and public administrative area)",
-        who_2019="Striking against or struck by other object(s) (School, other institution and public administrative area)",
-        clinically_equivalent=True,
-    ),
-    "W223": ReleaseTermDifference(
-        combined_2016="Striking against or struck by other objects (Sports and athletics area)",
-        who_2019="Striking against or struck by other object(s) (Sports and athletics area)",
-        clinically_equivalent=True,
-    ),
-    "W224": ReleaseTermDifference(
-        combined_2016="Striking against or struck by other objects (Street and highway)",
-        who_2019="Striking against or struck by other object(s) (Street and highway)",
-        clinically_equivalent=True,
-    ),
-    "W225": ReleaseTermDifference(
-        combined_2016="Striking against or struck by other objects (Trade and service area)",
-        who_2019="Striking against or struck by other object(s) (Trade and service area)",
-        clinically_equivalent=True,
-    ),
-    "W226": ReleaseTermDifference(
-        combined_2016="Striking against or struck by other objects (Industrial and construction area)",
-        who_2019="Striking against or struck by other object(s) (Industrial and construction area)",
-        clinically_equivalent=True,
-    ),
-    "W227": ReleaseTermDifference(
-        combined_2016="Striking against or struck by other objects (Farm)",
-        who_2019="Striking against or struck by other object(s) (Farm)",
-        clinically_equivalent=True,
-    ),
-    "W228": ReleaseTermDifference(
-        combined_2016="Striking against or struck by other objects (Other specified places)",
-        who_2019="Striking against or struck by other object(s) (Other specified places)",
-        clinically_equivalent=True,
-    ),
-    "W229": ReleaseTermDifference(
-        combined_2016="Striking against or struck by other objects (Unspecified place)",
-        who_2019="Striking against or struck by other object(s) (Unspecified place)",
-        clinically_equivalent=True,
-    ),
-    # W24: just the addition of "(s)" to "device" in the term, so clinically equivalent
-    "W24": ReleaseTermDifference(
-        combined_2016="Contact with lifting and transmission devices, not elsewhere classified",
-        who_2019="Contact with lifting and transmission device(s), not elsewhere classified",
-        clinically_equivalent=True,
-    ),
-    "W240": ReleaseTermDifference(
-        combined_2016="Contact with lifting and transmission devices, not elsewhere classified (Home)",
-        who_2019="Contact with lifting and transmission device(s), not elsewhere classified (Home)",
-        clinically_equivalent=True,
-    ),
-    "W241": ReleaseTermDifference(
-        combined_2016="Contact with lifting and transmission devices, not elsewhere classified (Residential institution)",
-        who_2019="Contact with lifting and transmission device(s), not elsewhere classified (Residential institution)",
-        clinically_equivalent=True,
-    ),
-    "W242": ReleaseTermDifference(
-        combined_2016="Contact with lifting and transmission devices, not elsewhere classified (School, other institution and public administrative area)",
-        who_2019="Contact with lifting and transmission device(s), not elsewhere classified (School, other institution and public administrative area)",
-        clinically_equivalent=True,
-    ),
-    "W243": ReleaseTermDifference(
-        combined_2016="Contact with lifting and transmission devices, not elsewhere classified (Sports and athletics area)",
-        who_2019="Contact with lifting and transmission device(s), not elsewhere classified (Sports and athletics area)",
-        clinically_equivalent=True,
-    ),
-    "W244": ReleaseTermDifference(
-        combined_2016="Contact with lifting and transmission devices, not elsewhere classified (Street and highway)",
-        who_2019="Contact with lifting and transmission device(s), not elsewhere classified (Street and highway)",
-        clinically_equivalent=True,
-    ),
-    "W245": ReleaseTermDifference(
-        combined_2016="Contact with lifting and transmission devices, not elsewhere classified (Trade and service area)",
-        who_2019="Contact with lifting and transmission device(s), not elsewhere classified (Trade and service area)",
-        clinically_equivalent=True,
-    ),
-    "W246": ReleaseTermDifference(
-        combined_2016="Contact with lifting and transmission devices, not elsewhere classified (Industrial and construction area)",
-        who_2019="Contact with lifting and transmission device(s), not elsewhere classified (Industrial and construction area)",
-        clinically_equivalent=True,
-    ),
-    "W247": ReleaseTermDifference(
-        combined_2016="Contact with lifting and transmission devices, not elsewhere classified (Farm)",
-        who_2019="Contact with lifting and transmission device(s), not elsewhere classified (Farm)",
-        clinically_equivalent=True,
-    ),
-    "W248": ReleaseTermDifference(
-        combined_2016="Contact with lifting and transmission devices, not elsewhere classified (Other specified places)",
-        who_2019="Contact with lifting and transmission device(s), not elsewhere classified (Other specified places)",
-        clinically_equivalent=True,
-    ),
-    "W249": ReleaseTermDifference(
-        combined_2016="Contact with lifting and transmission devices, not elsewhere classified (Unspecified place)",
-        who_2019="Contact with lifting and transmission device(s), not elsewhere classified (Unspecified place)",
-        clinically_equivalent=True,
-    ),
-    # Place of occurrence (W00–Y34)
-    # -----------------------------
-    # Per WHO ICD-10 coding guidance, a 4th character identifying place of occurrence
-    # must be assigned to all codes in categories W00–Y34.  The place digits come from
-    # modifier S20W00_4 (0=Home,... 9=Unspecified). There are two exceptions to this range:
-    #   Y06  Neglect and abandonment
-    #   Y07  Other maltreatment
-    #
-    # Originally all codes in the above range were 3 character codes, with no children.
-    # However, WHO introduced bespoke 4th-char subcategories for W26, X34, and X59 (in
-    # 2016 edition) so now the "4th character" place modifiers clash with the children
-    # of those codes. WHO guidance is that place of occurrence aren't actually 4th
-    # character modifiers, but should be stored elsewhere.
-    # However, UK clinical coding practice has chosen to ignore the WHO 4th char subcategories
-    # for W26/X34/X59 and continue to apply place modifiers to the 3-char parent codes.
-    #
-    # The activity modifier S20V01T_5 is intentionally ignored.
-    #
-    # Facts from https://classbrowser.nhs.uk/ref_books/ICD-10_2026_5th_Ed_NCCS.pdf
-    #
-    # 1. 4th character modifiers for W00-Y34
-    #   p217: A fourth character must be assigned with codes from categories W00-Y34
-    #         to identify where the injury, poisoning or adverse effect took place.
-    #         The fourth characters can be found in the ‘Place of occurrence code’
-    #         section at the beginning of the chapter. The exceptions are codes in
-    #         categories Y06.- Neglect and abandonment and Y07.- Other maltreatment.
-    #   p219: ICD-10 provides an activity subclassification as an extra character for
-    #         use with categories V01–Y34 to indicate the activity of the injured
-    #         person at the time the event occurred. However, due to the general
-    #         unavailability of this information, these activity subclassification
-    #         codes shown at the beginning of this chapter must not be used."
-    #   p220: The fourth character codes printed at categories
-    #             W26.- Contact with other sharp object(s)
-    #             X34.- Victim of earthquake and
-    #             X59.- Exposure to unspecified factor
-    #         in the ICD-10 Tabular List must not be used and must be crossed through
-    #         in the ICD-10 5th Edition books. The content that must be crossed
-    #         through can be found in the ICD-10 and OPCS-4 Classifications Content
-    #         Changes document on Delen. The ‘Place of occurrence codes’ must be used
-    #         for fourth character code assignment with categories W26, X34 and X59.
-    #
-    # Therefore the following codes are flagged as not clinically equivalent
-    "W260": ReleaseTermDifference(
-        combined_2016="Contact with other sharp object(s) (Home)",
-        who_2019="Contact with knife, sword or dagger",
-        clinically_equivalent=False,
-    ),
-    "W268": ReleaseTermDifference(
-        combined_2016="Contact with other sharp object(s) (Other specified places)",
-        who_2019="Contact with other sharp object(s), not elsewhere classified",
-        clinically_equivalent=False,
-    ),
-    "W269": ReleaseTermDifference(
-        combined_2016="Contact with other sharp object(s) (Unspecified place)",
-        who_2019="Contact with unspecified sharp object(s)",
-        clinically_equivalent=False,
-    ),
-    "X340": ReleaseTermDifference(
-        combined_2016="Victim of earthquake (Home)",
-        who_2019="Victim of cataclysmic earth movements caused by earthquake",
-        clinically_equivalent=False,
-    ),
-    "X341": ReleaseTermDifference(
-        combined_2016="Victim of earthquake (Residential institution)",
-        who_2019="Victim of tsunami",
-        clinically_equivalent=False,
-    ),
-    "X348": ReleaseTermDifference(
-        combined_2016="Victim of earthquake (Other specified places)",
-        who_2019="Victim of other specified effects of earthquake",
-        clinically_equivalent=False,
-    ),
-    "X349": ReleaseTermDifference(
-        combined_2016="Victim of earthquake (Unspecified place)",
-        who_2019="Victim of unspecified effect of earthquake",
-        clinically_equivalent=False,
-    ),
-    "X590": ReleaseTermDifference(
-        combined_2016="Exposure to unspecified factor (Home)",
-        who_2019="Exposure to unspecified factor causing fracture",
-        clinically_equivalent=False,
-    ),
-    "X599": ReleaseTermDifference(
-        combined_2016="Exposure to unspecified factor (Unspecified place)",
-        who_2019="Exposure to unspecified factor causing other and unspecified injury",
-        clinically_equivalent=False,
-    ),
-    # X47 in 2016 has an inclusion note for "carbon monoxide" so the addition of this to the term
-    # has not changed the code and we can class them as clinically equivalent
-    "X47": ReleaseTermDifference(
-        combined_2016="Accidental poisoning by and exposure to other gases and vapours",
-        who_2019="Accidental poisoning by and exposure to carbon monoxide and other gases and vapours",
-        clinically_equivalent=True,
-    ),
-    # X470 in 2016 (modifier of X47) conflicts with the 2019 X470 as a child code of X47 - not clinically equivalent
-    "X470": ReleaseTermDifference(
-        combined_2016="Accidental poisoning by and exposure to other gases and vapours (Home)",
-        who_2019="Accidental poisoning by and exposure to carbon monoxide from combustion engine exhaust",
-        clinically_equivalent=False,
-    ),
-    # X471 in 2016 (modifier of X47) conflicts with the 2019 X471 as a child code of X47 - not clinically equivalent
-    "X471": ReleaseTermDifference(
-        combined_2016="Accidental poisoning by and exposure to other gases and vapours (Residential institution)",
-        who_2019="Accidental poisoning by and exposure to carbon monoxide from utility gas",
-        clinically_equivalent=False,
-    ),
-    # X472 in 2016 (modifier of X47) conflicts with the 2019 X472 as a child code of X47 - not clinically equivalent
-    "X472": ReleaseTermDifference(
-        combined_2016="Accidental poisoning by and exposure to other gases and vapours (School, other institution and public administrative area)",
-        who_2019="Accidental poisoning by and exposure to carbon monoxide from other domestic fuels",
-        clinically_equivalent=False,
-    ),
-    # X473 in 2016 (modifier of X47) conflicts with the 2019 X473 as a child code of X47 - not clinically equivalent
-    "X473": ReleaseTermDifference(
-        combined_2016="Accidental poisoning by and exposure to other gases and vapours (Sports and athletics area)",
-        who_2019="Accidental poisoning by and exposure to carbon monoxide from other sources",
-        clinically_equivalent=False,
-    ),
-    # X474 in 2016 (modifier of X47) conflicts with the 2019 X474 as a child code of X47 - not clinically equivalent
-    "X474": ReleaseTermDifference(
-        combined_2016="Accidental poisoning by and exposure to other gases and vapours (Street and highway)",
-        who_2019="Accidental poisoning by and exposure to carbon monoxide from unspecified sources",
-        clinically_equivalent=False,
-    ),
-    "X475": ReleaseTermDifference(
-        combined_2016="Accidental poisoning by and exposure to other gases and vapours (Trade and service area)",
-        who_2019="Accidental poisoning by and exposure to carbon monoxide and other gases and vapours (Trade and service area)",
-        clinically_equivalent=True,
-    ),
-    "X476": ReleaseTermDifference(
-        combined_2016="Accidental poisoning by and exposure to other gases and vapours (Industrial and construction area)",
-        who_2019="Accidental poisoning by and exposure to carbon monoxide and other gases and vapours (Industrial and construction area)",
-        clinically_equivalent=True,
-    ),
-    "X477": ReleaseTermDifference(
-        combined_2016="Accidental poisoning by and exposure to other gases and vapours (Farm)",
-        who_2019="Accidental poisoning by and exposure to carbon monoxide and other gases and vapours (Farm)",
-        clinically_equivalent=True,
-    ),
-    # X478 in 2016 (modifier of X47) conflicts with the 2019 X478 as a child code of X47 - not clinically equivalent
-    "X478": ReleaseTermDifference(
-        combined_2016="Accidental poisoning by and exposure to other gases and vapours (Other specified places)",
-        who_2019="Accidental poisoning by and exposure to other specified gases and vapours",
-        clinically_equivalent=False,
-    ),
-    # X479 in 2016 (modifier of X47) conflicts with the 2019 X479 as a child code of X47 - not clinically equivalent
-    "X479": ReleaseTermDifference(
-        combined_2016="Accidental poisoning by and exposure to other gases and vapours (Unspecified place)",
-        who_2019="Accidental poisoning by and exposure to unspecified gases and vapours",
-        clinically_equivalent=False,
-    ),
-    # X67 in 2016 has an inclusion note for "carbon monoxide" so the addition of this to the term
-    # has not changed the code and we can class them as clinically equivalent
-    "X67": ReleaseTermDifference(
-        combined_2016="Intentional self-poisoning by and exposure to other gases and vapours",
-        who_2019="Intentional self-poisoning by and exposure to carbon monoxide and other gases and vapours",
-        clinically_equivalent=True,
-    ),
-    # X670 in 2016 (modifier of X67) conflicts with the 2019 X670 as a child code of X67 - not clinically equivalent
-    "X670": ReleaseTermDifference(
-        combined_2016="Intentional self-poisoning by and exposure to other gases and vapours (Home)",
-        who_2019="Intentional self-poisoning by and exposure to carbon monoxide from combustion engine exhaust",
-        clinically_equivalent=False,
-    ),
-    # X671 in 2016 (modifier of X67) conflicts with the 2019 X671 as a child code of X67 - not clinically equivalent
-    "X671": ReleaseTermDifference(
-        combined_2016="Intentional self-poisoning by and exposure to other gases and vapours (Residential institution)",
-        who_2019="Intentional self-poisoning by and exposure to carbon monoxide from utility gas",
-        clinically_equivalent=False,
-    ),
-    # X672 in 2016 (modifier of X67) conflicts with the 2019 X672 as a child code of X67 - not clinically equivalent
-    "X672": ReleaseTermDifference(
-        combined_2016="Intentional self-poisoning by and exposure to other gases and vapours (School, other institution and public administrative area)",
-        who_2019="Intentional self-poisoning by and exposure to carbon monoxide from other domestic fuels",
-        clinically_equivalent=False,
-    ),
-    # X673 in 2016 (modifier of X67) conflicts with the 2019 X673 as a child code of X67 - not clinically equivalent
-    "X673": ReleaseTermDifference(
-        combined_2016="Intentional self-poisoning by and exposure to other gases and vapours (Sports and athletics area)",
-        who_2019="Intentional self-poisoning by and exposure to carbon monoxide from other sources",
-        clinically_equivalent=False,
-    ),
-    # X674 in 2016 (modifier of X67) conflicts with the 2019 X674 as a child code of X67 - not clinically equivalent
-    "X674": ReleaseTermDifference(
-        combined_2016="Intentional self-poisoning by and exposure to other gases and vapours (Street and highway)",
-        who_2019="Intentional self-poisoning by and exposure to carbon monoxide from unspecified sources",
-        clinically_equivalent=False,
-    ),
-    "X675": ReleaseTermDifference(
-        combined_2016="Intentional self-poisoning by and exposure to other gases and vapours (Trade and service area)",
-        who_2019="Intentional self-poisoning by and exposure to carbon monoxide and other gases and vapours (Trade and service area)",
-        clinically_equivalent=True,
-    ),
-    "X676": ReleaseTermDifference(
-        combined_2016="Intentional self-poisoning by and exposure to other gases and vapours (Industrial and construction area)",
-        who_2019="Intentional self-poisoning by and exposure to carbon monoxide and other gases and vapours (Industrial and construction area)",
-        clinically_equivalent=True,
-    ),
-    "X677": ReleaseTermDifference(
-        combined_2016="Intentional self-poisoning by and exposure to other gases and vapours (Farm)",
-        who_2019="Intentional self-poisoning by and exposure to carbon monoxide and other gases and vapours (Farm)",
-        clinically_equivalent=True,
-    ),
-    # X678 in 2016 (modifier of X67) conflicts with the 2019 X678 as a child code of X67 - not clinically equivalent
-    "X678": ReleaseTermDifference(
-        combined_2016="Intentional self-poisoning by and exposure to other gases and vapours (Other specified places)",
-        who_2019="Intentional self-poisoning by and exposure to other specified gases and vapours",
-        clinically_equivalent=False,
-    ),
-    # X679 in 2016 (modifier of X67) conflicts with the 2019 X679 as a child code of X67 - not clinically equivalent
-    "X679": ReleaseTermDifference(
-        combined_2016="Intentional self-poisoning by and exposure to other gases and vapours (Unspecified place)",
-        who_2019="Intentional self-poisoning by and exposure to unspecified gases and vapours",
-        clinically_equivalent=False,
-    ),
-    # X880 in 2016 (modifier of X88) conflicts with the 2019 X880 as a child code of X88 - not clinically equivalent
-    "X880": ReleaseTermDifference(
-        combined_2016="Assault by gases and vapours (Home)",
-        who_2019="Assault by carbon monoxide from combustion engine exhaust",
-        clinically_equivalent=False,
-    ),
-    # X881 in 2016 (modifier of X88) conflicts with the 2019 X881 as a child code of X88 - not clinically equivalent
-    "X881": ReleaseTermDifference(
-        combined_2016="Assault by gases and vapours (Residential institution)",
-        who_2019="Assault by carbon monoxide from utility gas",
-        clinically_equivalent=False,
-    ),
-    # X882 in 2016 (modifier of X88) conflicts with the 2019 X882 as a child code of X88 - not clinically equivalent
-    "X882": ReleaseTermDifference(
-        combined_2016="Assault by gases and vapours (School, other institution and public administrative area)",
-        who_2019="Assault by carbon monoxide from other domestic fuels",
-        clinically_equivalent=False,
-    ),
-    # X883 in 2016 (modifier of X88) conflicts with the 2019 X883 as a child code of X88 - not clinically equivalent
-    "X883": ReleaseTermDifference(
-        combined_2016="Assault by gases and vapours (Sports and athletics area)",
-        who_2019="Assault by carbon monoxide from other sources",
-        clinically_equivalent=False,
-    ),
-    # X884 in 2016 (modifier of X88) conflicts with the 2019 X884 as a child code of X88 - not clinically equivalent
-    "X884": ReleaseTermDifference(
-        combined_2016="Assault by gases and vapours (Street and highway)",
-        who_2019="Assault by carbon monoxide from unspecified sources",
-        clinically_equivalent=False,
-    ),
-    # X888 in 2016 (modifier of X88) conflicts with the 2019 X888 as a child code of X88 - not clinically equivalent
-    "X888": ReleaseTermDifference(
-        combined_2016="Assault by gases and vapours (Other specified places)",
-        who_2019="Assault by other specified gases and vapours",
-        clinically_equivalent=False,
-    ),
-    # X889 in 2016 (modifier of X88) conflicts with the 2019 X889 as a child code of X88 - not clinically equivalent
-    "X889": ReleaseTermDifference(
-        combined_2016="Assault by gases and vapours (Unspecified place)",
-        who_2019="Assault by unspecified gases and vapours",
-        clinically_equivalent=False,
-    ),
-    # Y17 in 2016 has an inclusion note for "carbon monoxide" so the addition of this to the term
-    # has not changed the code and we can class them as clinically equivalent
-    "Y17": ReleaseTermDifference(
-        combined_2016="Poisoning by and exposure to other gases and vapours, undetermined intent",
-        who_2019="Poisoning by and exposure to carbon monoxide and other gases and vapours, undetermined intent",
-        clinically_equivalent=True,
-    ),
-    # Y170 in 2016 (modifier of Y17) conflicts with the 2019 Y170 as a child code of Y17 - not clinically equivalent
-    "Y170": ReleaseTermDifference(
-        combined_2016="Poisoning by and exposure to other gases and vapours, undetermined intent (Home)",
-        who_2019="Poisoning by and exposure to carbon monoxide from combustion engine exhaust, undetermined intent",
-        clinically_equivalent=False,
-    ),
-    # Y171 in 2016 (modifier of Y17) conflicts with the 2019 Y171 as a child code of Y17 - not clinically equivalent
-    "Y171": ReleaseTermDifference(
-        combined_2016="Poisoning by and exposure to other gases and vapours, undetermined intent (Residential institution)",
-        who_2019="Poisoning by and exposure to carbon monoxide from utility gas, undetermined intent",
-        clinically_equivalent=False,
-    ),
-    # Y172 in 2016 (modifier of Y17) conflicts with the 2019 Y172 as a child code of Y17 - not clinically equivalent
-    "Y172": ReleaseTermDifference(
-        combined_2016="Poisoning by and exposure to other gases and vapours, undetermined intent (School, other institution and public administrative area)",
-        who_2019="Poisoning by and exposure to carbon monoxide from other domestic fuels, undetermined intent",
-        clinically_equivalent=False,
-    ),
-    # Y173 in 2016 (modifier of Y17) conflicts with the 2019 Y173 as a child code of Y17 - not clinically equivalent
-    "Y173": ReleaseTermDifference(
-        combined_2016="Poisoning by and exposure to other gases and vapours, undetermined intent (Sports and athletics area)",
-        who_2019="Poisoning by and exposure to carbon monoxide from other sources, undetermined intent",
-        clinically_equivalent=False,
-    ),
-    # Y174 in 2016 (modifier of Y17) conflicts with the 2019 Y174 as a child code of Y17 - not clinically equivalent
-    "Y174": ReleaseTermDifference(
-        combined_2016="Poisoning by and exposure to other gases and vapours, undetermined intent (Street and highway)",
-        who_2019="Poisoning by and exposure to carbon monoxide from unspecified sources, undetermined intent",
-        clinically_equivalent=False,
-    ),
-    "Y175": ReleaseTermDifference(
-        combined_2016="Poisoning by and exposure to other gases and vapours, undetermined intent (Trade and service area)",
-        who_2019="Poisoning by and exposure to carbon monoxide and other gases and vapours, undetermined intent (Trade and service area)",
-        clinically_equivalent=True,
-    ),
-    "Y176": ReleaseTermDifference(
-        combined_2016="Poisoning by and exposure to other gases and vapours, undetermined intent (Industrial and construction area)",
-        who_2019="Poisoning by and exposure to carbon monoxide and other gases and vapours, undetermined intent (Industrial and construction area)",
-        clinically_equivalent=True,
-    ),
-    "Y177": ReleaseTermDifference(
-        combined_2016="Poisoning by and exposure to other gases and vapours, undetermined intent (Farm)",
-        who_2019="Poisoning by and exposure to carbon monoxide and other gases and vapours, undetermined intent (Farm)",
-        clinically_equivalent=True,
-    ),
-    # Y178 in 2016 (modifier of Y17) conflicts with the 2019 Y178 as a child code of Y17 - not clinically equivalent
-    "Y178": ReleaseTermDifference(
-        combined_2016="Poisoning by and exposure to other gases and vapours, undetermined intent (Other specified places)",
-        who_2019="Poisoning by and exposure to other specified gases and vapours, undetermined intent",
-        clinically_equivalent=False,
-    ),
-    # Y179 in 2016 (modifier of Y17) conflicts with the 2019 Y179 as a child code of Y17 - not clinically equivalent
-    "Y179": ReleaseTermDifference(
-        combined_2016="Poisoning by and exposure to other gases and vapours, undetermined intent (Unspecified place)",
-        who_2019="Poisoning by and exposure to unspecified gases and vapours, undetermined intent",
-        clinically_equivalent=False,
-    ),
-    # The 2016 inclusion note makes clear this is just for evaluations that turn out to be negative
-    # so these are clinically equivalent
-    "Z03": ReleaseTermDifference(
-        combined_2016="Medical observation and evaluation for suspected diseases and conditions",
-        who_2019="Medical observation and evaluation for suspected diseases and conditions, ruled out",
-        clinically_equivalent=True,
-    ),
-}
-
-
-def is_2016_claml_only(code: str) -> bool:
-    """Check if a code is expected to be only in the 2016 CLAML data."""
-    return code in CLAML_VS_SCRAPED_DIFFERENCES.claml_only
-
-
-def is_2016_scraped_only(code: str) -> bool:
-    """Check if a code is expected to be only in the 2016 scraped data."""
-    return code in CLAML_VS_SCRAPED_DIFFERENCES.scraped_only
-
-
-def is_2016_description_difference(
-    code: str,
-    claml_description: str,
-    scraped_description: str,
-) -> bool:
-    """Check if a code is expected to have a different description in the 2016 CLAML vs scraped data."""
-    known = CLAML_VS_SCRAPED_DIFFERENCES.term_differences.get(code)
-    return (
-        known is not None
-        and known.claml == claml_description
-        and known.scraped == scraped_description
-    )
-
-
-def should_include_2016_claml_only(code: str) -> bool:
-    """For a ClAML-only code, check if it should be included in the release."""
-    known = CLAML_VS_SCRAPED_DIFFERENCES.claml_only.get(code)
-    return known is not None and known.include_in_release
-
-
-def should_include_2016_scraped_only(code: str) -> bool:
-    """For a scraped-only code, check if it should be included in the release."""
-    known = CLAML_VS_SCRAPED_DIFFERENCES.scraped_only.get(code)
-    return known is not None and known.include_in_release
-
-
-def should_use_scraped_for_2016(code: str) -> bool:
-    """For a code with a description difference, should we use the scraped description?"""
-    known = CLAML_VS_SCRAPED_DIFFERENCES.term_differences.get(code)
-    return known is not None and known.use == "scraped"
-
-
-def get_2016_2019_description_difference(
-    code: str,
-    combined_2016_description: str,
-    who_2019_description: str,
-) -> ReleaseTermDifference | None:
-    """
-    Check if a code is expected to have a different description and return the known
-    difference if so. Including the boolean for whether the descriptions are clinically
-    equivalent or not.
-    """
-    known = COMBINED_2016_VS_2019_DIFFERENCES.get(code)
-    if known is None:
-        return None
-    if (
-        known.combined_2016 == combined_2016_description
-        and known.who_2019 == who_2019_description
-    ):
-        return known
-    return None
-
-
-def clinically_different_codes(codes: list[str]) -> dict[str, dict[str, str]]:
-    """
-    Given a list of codes, return a dict containing the codes
-    that have clinically different descriptions between the 2016 and 2019
-    releases, along with their descriptions in both releases.
-    """
-    differences = {}
-    normalised_codes = set(code.upper() for code in codes)
-    for code in normalised_codes:
-        difference = COMBINED_2016_VS_2019_DIFFERENCES.get(code)
-        if difference and not difference.clinically_equivalent:
-            differences[code] = {
-                "combined_2016": difference.combined_2016,
-                "who_2019": difference.who_2019,
-            }
-    return differences
-
-
-def codes_with_different_descriptions(
-    codes: list[str],
-) -> dict[str, dict[str, str | bool]]:
-    """
-    Given a list of codes, return a dict containing the codes
-    that have different descriptions between the 2016 and 2019
-    releases, along with their descriptions in both releases.
-    """
-    differences = {}
-    normalised_codes = set(code.upper() for code in codes)
-    for code in normalised_codes:
-        difference = COMBINED_2016_VS_2019_DIFFERENCES.get(code)
-        if difference:
-            differences[code] = {
-                "combined_2016": difference.combined_2016,
-                "who_2019": difference.who_2019,
-                "equivalent": difference.clinically_equivalent,
-            }
-    return differences
-
-
-# Moved code sets. We flag codelists that contain only part of one of these sets,
-# because the same clinical concept is represented by different codes across releases.
-# These moved codes were derived from this report (https://github.com/bennettoxford/icd-browser-scraper/blob/main/claml/2016-vs-2019.md)
-# which shows all the codes that appear in one set and not the other.
-MOVED_CODE_SETS = (
-    {
-        "title": "Pneumocystosis",
-        "nhs2016": ["B59"],
-        "who2019": ["B485"],
-        "comment": "This is B59 in 2016, but B485 in 2019.",
-    },
-    {
-        "title": "Causalgia",
-        "nhs2016": ["G564"],
-        "who2019": ["G906"],
-        "comment": "This is G564 in 2016, but G906 in 2019.",
-    },
-    {
-        "title": "Irritable bowel syndrome",
-        "nhs2016": ["K58", "K580", "K589"],
-        "who2019": ["K58", "K581", "K582", "K583", "K588"],
-        "comment": "The codes for this were K580 and K589 in 2016, but K581, K582, K583 and K588 in 2019. You likely want all these codes in your codelist. However if you are specifically looking for diarrhoea, or constipation, rather than IBS, then you may only want some but not all of these codes.",
-    },
-    {
-        "title": "Zika virus disease",
-        "nhs2016": ["U06", "U069"],
-        "who2019": ["A925"],
-        "comment": "This is U06/U069 in 2016, but is A925 in 2019.",
-    },
-    {
-        "title": "Personal history of COVID-19",
-        "nhs2016": ["U073"],
-        "who2019": ["U08", "U089"],
-        "comment": "This is U073 in 2016, but U08/U089 in 2019.",
-    },
-    {
-        "title": "Post COVID-19 condition",
-        "nhs2016": ["U074"],
-        "who2019": ["U09", "U099"],
-        "comment": "This is U074 in 2016, but U09/U099 in 2019.",
-    },
-    {
-        "title": "Multisystem inflammatory syndrome associated with COVID-19",
-        "nhs2016": ["U075"],
-        "who2019": ["U10", "U109"],
-        "comment": "This is U075 in 2016, but U10/U109 in 2019.",
-    },
-    {
-        "title": "Need for immunization against COVID-19",
-        "nhs2016": ["U076"],
-        "who2019": ["U11", "U119"],
-        "comment": "This is U076 in 2016, but U11/U119 in 2019.",
-    },
-    {
-        "title": "COVID-19 vaccines causing adverse effects in therapeutic use",
-        "nhs2016": ["U077"],
-        "who2019": ["U12", "U129"],
-        "comment": "This is U077 in 2016, but U12/U129 in 2019.",
-    },
-)
-
-
-def moved_codes(codes: list[str]) -> list[dict]:
-    """
-    Given a list of codes, return a list of moved code sets that contain any of the
-    codes. Each moved code set contains the title, the codes in both releases, and a
-    comment explaining the change.
-    """
-    normalised_codes = set(code.upper() for code in codes)
-    moved_codes = []
-    for moved_code_set in MOVED_CODE_SETS:
-        possible_codes = moved_code_set["nhs2016"] + moved_code_set["who2019"]
-        if any(code in normalised_codes for code in possible_codes):
-            moved_codes.append(moved_code_set)
-    return moved_codes
-
-
-@dataclass
-class RubricChange:
-    """
-    Represents a change to the rubrics of a code between releases. This can
-    include additions, removals, and replacements of rubric text. The `who_2016`
-    field represents the original rubrics from the 2016 WHO release, while the
-    `remove`, `add`, and `replace` fields represent the changes to be applied.
-    """
-
-    who_2016: dict[str, list[str]] = field(default_factory=dict)
-    remove: dict[str, list[str]] = field(default_factory=dict)
-    add: dict[str, list[str]] = field(default_factory=dict)
-    replace: dict[str, dict[str, str]] = field(default_factory=dict)
-    comment: str = ""
-
-    @property
-    def resolved_rubrics(self) -> dict[str, list[str]]:
-        """
-        Resolve the rubric changes by applying removals, additions, and replacements
-        to the original WHO rubrics. Returns a dictionary of rubric types to their
-        final list of values after applying the changes.
-        """
-        rubrics = {
-            # Using list() to create a shallow copy of the list to avoid mutating
-            # the original WHO rubrics
-            rubric_type: list(values)
-            for rubric_type, values in self.who_2016.items()
-        }
-
-        for rubric_type, replacements in self.replace.items():
-            for text_to_find, text_to_replace in replacements.items():
-                rubrics[rubric_type] = [
-                    v.replace(text_to_find, text_to_replace)
-                    for v in rubrics[rubric_type]
-                ]
-
-        for rubric_type, values_to_remove in self.remove.items():
-            values = rubrics[rubric_type]
-            for value in values_to_remove:
-                values.remove(value)
-            if not values:
-                del rubrics[rubric_type]
-
-        for rubric_type, values_to_add in self.add.items():
-            rubrics.setdefault(rubric_type, []).extend(values_to_add)
-
-        return rubrics
-
-
-KNOWN_2016_RUBRIC_CHANGES = {
-    # Things where the NHS has corrected something - and it appears in the pdf of ICD-10_Classification_Content_Changes_2026_.pdf
-    "B81": RubricChange(
-        who_2016={
-            "exclusion": [
-                "Angiostrongyliasis due to: Angiostrongylus costaricensis (B83.2)",
-                "Angiostrongyliasis due to: Parastrongylus costaricensis (B83.2)",
-            ]
-        },
-        replace={"exclusion": {"costaricensis": "cantonensis"}},
-        comment="Typo in WHO. B81 is about _costaricensis_, so _cantonensis_ not _costaricensis_ should be excluded. Correction mentioned on p10 of ICD-10_Classification_Content_Changes_2026_.pdf",
-    ),
-    "X54": RubricChange(
-        who_2016={
-            "inclusion": [
-                "lack of water as the cause of: dehydration",
-                "lack of water as the cause of: inanition",
-            ],
-            "exclusion": [
-                "insufficient intake of water due to self neglect (R63.3)",
-                "neglect or abandonment by others (Y06.-)",
-                "self neglect NOS (R46.8)",
-            ],
-        },
-        replace={"exclusion": {"R63.3": "R63.6"}},
-        comment="Typo in WHO. Link should be to R63.6 and not R63.3. Correction mentioned on p12 of ICD-10_Classification_Content_Changes_2026_.pdf",
-    ),
-    "K296": RubricChange(
-        who_2016={
-            "inclusion": [
-                "Giant hypertrophic gastritis",
-                "Granulomatous gastritis",
-                "Ménétrier disease",
-            ],
-            "exclusion": [
-                "with gastro-oesophageal reflux disease (K21.-)",
-                "with Helicobacter pylori associated chronic gastritis (K29.5)",
-            ],
-        },
-        remove={
-            "exclusion": [
-                "with gastro-oesophageal reflux disease (K21.-)",
-                "with Helicobacter pylori associated chronic gastritis (K29.5)",
-            ],
-        },
-        comment="NHS removes the two exclusion rubrics associated with this code. Correction mentioned on p12 of ICD-10_Classification_Content_Changes_2026_.pdf",
-    ),
-    "N281": RubricChange(
-        who_2016={
-            "inclusion": ["Cyst (acquired)(multiple)(solitary) of kidney"],
-            "exclusion": ["cystic kidney disease (congenital) (Q61.-)"],
-        },
-        replace={"exclusion": {" (congenital)": ", congenital"}},
-        comment="Change '(congenital)' to ', congenital'. Correction mentioned on p11 of ICD-10_Classification_Content_Changes_2026_.pdf",
-    ),
-    # Things where the NHS has corrected a typo
-    "F432": RubricChange(
-        who_2016={
-            "definition": [
-                "States of subjective distress and emotional disturbance, usually interfering with social functioning and performance, arising in the period of adaptation to a significant life change or a stressful life event. The stressor may have affected the integrity of an individual's social network (bereavement, separation experiences) or the wider system of social supports and values (migration, refugee status), or represented a major developmental transition or crisis (going to school, becoming a parent, failure to attain a cherished personal goal, retirement). Individual predisposition or vulnerability plays an important role in the risk of occurrence and the shaping of the manifestations of adjustment disorders, but it is nevertheless assumed that the condition would not have arisen without the stressor. The manifestations vary and include depressed mood, anxiety or worry (or mixture of these), a feeling of inability to cope, plan ahead, or continue in the present situation, as well as some degree of disability in 9the performance of daily routine. Conduct disorders may be an associated feature, particularly in adolescents. The predominant feature may be a brief or prolonged depressive reaction, or a disturbance of other emotions and conduct."
-            ],
-            "inclusion": ["Culture shock", "Grief reaction", "Hospitalism in children"],
-            "exclusion": ["separation anxiety disorder of childhood (F93.0)"],
-        },
-        replace={"definition": {"9the": "the"}},
-        comment="'9the' in WHO becomes 'the' in NHS",
-    ),
-    "S00": RubricChange(
-        who_2016={
-            "exclusion": [
-                "cerebral contusion (diffuse) (S06.2)",
-                "focal cerebral contusion (diffuse) (S06.3)",
-                "injury of eye and orbit (S05.-)",
-            ]
-        },
-        comment="The NHS classification browser lists '• focal (S06.3)' underneath cerebral contusion (diffuse). Therefore this is a formatting difference. We use the WHO rubric as-is.",
-    ),
-    "M544": RubricChange(
-        who_2016={"exclusion": ["that due to intervertebral disc disorder (M51.1)"]},
-        replace={"exclusion": {"that due to": "due to"}},
-        comment="NHS removes the non-idiomatic 'that'",
-    ),
-    "M248": RubricChange(
-        who_2016={"exclusion": ["that involving iliotibial band syndrome (M76.3)"]},
-        replace={"exclusion": {"that involving": "involving"}},
-        comment="NHS removes the non-idiomatic 'that'",
-    ),
-    "Z21": RubricChange(
-        who_2016={
-            "inclusion": ["HIV positive NOS"],
-            "exclusion": [
-                "contact with or exposure to human immunodeficiency virus [HIV] (Z20.6)",
-                "human immunodeficiency virus [HIV] disease (B20-B24)",
-                "lhuman immunodeficiency virus [HIV] disease complicating pregnancy, childbirth and the puerperium (O98.7)",
-                "laboratory evidence of human immunodeficiency virus [HIV] (R75)",
-            ],
-        },
-        replace={"exclusion": {"lhuman": "human"}},
-        comment="'lhuman' in WHO becomes 'human' in NHS",
-    ),
-    "O432": RubricChange(
-        who_2016={
-            "coding-hint": [
-                "Use additional code, if desired, to identify any: postpartum haemmorrhage, third stage (O72.0)",
-                "Use additional code, if desired, to identify any: retained placenta without hemorrhage (O73.0)",
-            ],
-            "inclusion": [
-                "Placenta: accreta",
-                "Placenta: increta",
-                "Placenta: percreta",
-            ],
-        },
-        replace={"coding-hint": {"haemmorrhage": "haemorrhage"}},
-        comment="'haemmorrhage' in WHO becomes 'haemorrhage' in NHS - but NB hemorrhage stays and is not replaced with haemorrhage",
-    ),
-    "V98": RubricChange(
-        who_2016={
-            "inclusion": [
-                "accident to, on or involving: cable-car, not on rails",
-                "accident to, on or involving: ice-yacht",
-                "accident to, on or involving: land-yacht",
-                "accident to, on or involving: ski chair-lift",
-                "accident to, on or involving: ski-lift with gondola",
-                "caught or dragged by cable-car, not on rails",
-                "fall or jump from cable-car, not on rails",
-                "object thrown from or in cable-car, not on rails",
-            ]
-        },
-        comment="WHO browser omits the final 3 'cable-car, not on rails' - but the claml includes it, so can use the WHO value as-is",
-    ),
-    "O96": RubricChange(
-        who_2016={
-            "note": [
-                "This category is to be used to indicate death from any obstetric cause (conditions in categories O00-O75O85-O92, andO98-O99 occurring more than 42 days but less than one year after delivery."
-            ],
-            "coding-hint": [
-                "Use additional code, if desired, to identify obstetric cause (direct or indirect) of death."
-            ],
-            "exclusion": [
-                "death from conditions specified as sequelae or late effects of obstetric causes (O97.-)",
-                "conditions specified as sequelae or late effects of obstetric causes that are not resulting in death (O94)",
-            ],
-        },
-        replace={
-            "note": {
-                "(conditions in categories O00-O75O85-O92, andO98-O99 occurring": "(conditions in categories O00-O75, O85-O92, and O98-O99) occurring"
-            },
-        },
-        comment="NHS better formatting of the note - adds commas, spaces and the missing closing parenthesis",
-    ),
-    "X36": RubricChange(
-        who_2016={
-            "inclusion": ["mudslide of cataclysmic nature"],
-            "exclusion": [
-                "earthquake (X34.-)",
-                "transport accident involving collision with avalanche or landslide not in motion (V01-V99)",
-            ],
-        },
-        replace={"exclusion": {"X34.-": "X34"}},
-        comment="X34 has no children in the NHS browser so code link is X34 rather than X34.-",
-    ),
-    "X39": RubricChange(
-        who_2016={
-            "inclusion": ["natural radiation NOS", "tidal wave NOS"],
-            "exclusion": ["exposure NOS (X59.9)", "tsunami (X34.1)"],
-        },
-        replace={"exclusion": {"X34.1": "X34", "X59.9": "X59"}},
-        comment="X34 and X59 have no children in the NHS browser so code link is X34 and X59 rather than X34.1 and X59.9",
-    ),
-    "W26": RubricChange(
-        who_2016={},
-        add={
-            "exclusion": ["sharp object(s) embedded in skin ( W45 )"],
-        },
-        comment="NHS adds an exclusion that presumably WHO removed when they added child codes for this",
-    ),
-    "W45": RubricChange(
-        who_2016={
-            "inclusion": ["edge of stiff paper", "nail", "splinter", "tin-can lid"],
-            "exclusion": [
-                "contact with: hand tools (nonpowered)(powered) (W27-W29)",
-                "contact with: hypodermic needle (W46)",
-                "contact with: other sharp objects (W26.-)",
-                "contact with: sharp glass (W25)",
-                "struck by objects (W20-W22)",
-            ],
-        },
-        replace={"exclusion": {"other sharp objects": "other sharp object(s)"}},
-        remove={
-            "inclusion": ["edge of stiff paper", "tin-can lid"],
-        },
-        comment="WHO has the extra examples of stiff paper and tin-can lid. But this is likely a mistake because those two items fall as examples under W26 which seems a better place for them",
-    ),
-    "R296": RubricChange(
-        who_2016={
-            "inclusion": [
-                "Tendency to fall because of old age or other unclear health problems"
-            ],
-            "exclusion": [
-                "accidents (X59.-)",
-                "difficulty in walking (R26.2)",
-                "dizziness and giddiness (R42)",
-                "falls causing injury (W00-W19)",
-                "falls due to diseases classified elsewhere",
-                "syncope and collapse (R55)",
-            ],
-        },
-        replace={"exclusion": {"X59.-": "X59"}},
-        comment="X59 has no children in the NHS browser so code link is X59 rather than X59.-",
-    ),
-    # Formatting differences
-    "K750": RubricChange(
-        who_2016={
-            "inclusion": [
-                "Hepatic abscess: NOS",
-                "Hepatic abscess: cholangitic",
-                "Hepatic abscess: haematogenic",
-                "Hepatic abscess: lymphogenic",
-                "Hepatic abscess: pylephlebitic",
-            ],
-            "exclusion": [
-                "amoebic liver abscess (A06.4), (K77.0)",
-                "cholangitis without liver abscess (K83.0)",
-                "pylephlebitis without liver abscess (K75.1)",
-            ],
-        },
-        comment="NHS browser formats '(A06.4), (K77.0)' as '( A06.4 , , K77.* )'. WHO looks better so we'll use that as-is",
-    ),
-    "I05": RubricChange(
-        who_2016={
-            "inclusion": [
-                "conditions classifiable to (I05.0) and (I05.2-I05.9), whether specified as rheumatic or not"
-            ],
-            "exclusion": ["when specified as nonrheumatic (I34.-)"],
-        },
-        comment="NHS has '(I05.0, and I05.2-I05.9)'. WHO is probably better so we'll use that as-is",
-    ),
-    # Morphology/behaviour code differences
-    "D18": RubricChange(
-        who_2016={
-            "inclusion": ["morphology codes M912-M917 with behaviour code /0"],
-            "exclusion": ["blue or pigmented naevus (D22.-)"],
-        },
-        remove={
-            "inclusion": ["morphology codes M912-M917 with behaviour code /0"],
-        },
-        comment="Morphology behaviour codes are not core ICD-10 and are not used in the NHS browser.",
-    ),
-    "D17": RubricChange(
-        who_2016={"inclusion": ["morphology codes M885-M888 with behaviour code /0"]},
-        remove={
-            "inclusion": ["morphology codes M885-M888 with behaviour code /0"],
-        },
-        comment="Morphology behaviour codes are not core ICD-10 and are not used in the NHS browser.",
-    ),
-    "D19": RubricChange(
-        who_2016={"inclusion": ["morphology code M905 with behaviour code /0"]},
-        remove={
-            "inclusion": ["morphology code M905 with behaviour code /0"],
-        },
-        comment="Morphology behaviour codes are not core ICD-10 and are not used in the NHS browser.",
-    ),
-    "D22": RubricChange(
-        who_2016={
-            "inclusion": [
-                "morphology codes M872-M879 with behaviour code /0",
-                "naevus: NOS",
-                "naevus: blue",
-                "naevus: hairy",
-                "naevus: pigmented",
-            ]
-        },
-        remove={
-            "inclusion": ["morphology codes M872-M879 with behaviour code /0"],
-        },
-        comment="Morphology behaviour codes are not core ICD-10 and are not used in the NHS browser.",
-    ),
-    "D25": RubricChange(
-        who_2016={
-            "inclusion": [
-                "benign neoplasms of uterus with morphology code M889 and behaviour code /0",
-                "fibromyoma of uterus",
-            ]
-        },
-        remove={
-            "inclusion": [
-                "benign neoplasms of uterus with morphology code M889 and behaviour code /0",
-            ]
-        },
-        comment="Morphology behaviour codes are not core ICD-10 and are not used in the NHS browser.",
-    ),
-    "C45": RubricChange(
-        who_2016={"inclusion": ["morphology code M905 with behaviour code /3"]},
-        remove={
-            "inclusion": ["morphology code M905 with behaviour code /3"],
-        },
-        comment="Morphology behaviour codes are not core ICD-10 and are not used in the NHS browser.",
-    ),
-    "C46": RubricChange(
-        who_2016={"inclusion": ["morphology code M9140 with behaviour code /3"]},
-        remove={
-            "inclusion": ["morphology code M9140 with behaviour code /3"],
-        },
-        comment="Morphology behaviour codes are not core ICD-10 and are not used in the NHS browser.",
-    ),
-    "C43": RubricChange(
-        who_2016={
-            "inclusion": ["morphology codes M872-M879 with behaviour code /3"],
-            "exclusion": [
-                "malignant melanoma of skin of genital organs (C51-C52) (C60.-) (C63.-)"
-            ],
-        },
-        remove={
-            "inclusion": ["morphology codes M872-M879 with behaviour code /3"],
-        },
-        comment="Morphology behaviour codes are not core ICD-10 and are not used in the NHS browser.",
-    ),
-    "D03": RubricChange(
-        who_2016={"inclusion": ["morphology codes M872-M879 with behaviour code /2"]},
-        remove={
-            "inclusion": ["morphology codes M872-M879 with behaviour code /2"],
-        },
-        comment="Morphology behaviour codes are not core ICD-10 and are not used in the NHS browser.",
-    ),
-    "H200": RubricChange(
-        who_2016={
-            "inclusion": [
-                "Anterior uveitis acute, recurrent or subacute",
-                "Cyclitis acute, recurrent or subacute",
-                "Iritis acute, recurrent or subacute",
-            ]
-        },
-        replace={
-            "inclusion": {"recurrent or subacute": "recurrent or subacute of cornea"}
-        },
-        comment="NHS adds 'of cornea' to the end of the line - seems a minimal difference",
-    ),
-    "J44": RubricChange(
-        who_2016={
-            "inclusion": [
-                "chronic: bronchitis: asthmatic (obstructive)",
-                "chronic: bronchitis: emphysematous",
-                "chronic: bronchitis: with: airways obstruction",
-                "chronic: bronchitis: with: emphysema",
-                "chronic: obstructive: asthma",
-                "chronic: obstructive: bronchitis",
-                "chronic: obstructive: tracheobronchitis",
-            ],
-            "exclusion": [
-                "asthma (J45.-)",
-                "asthmatic bronchitis NOS (J45.9)",
-                "bronchiectasis (J47)",
-                "chronic: bronchitis: NOS (J42)",
-                "chronic: bronchitis: simple and mucopurulent (J41.-)",
-                "chronic: tracheitis (J42)",
-                "chronic: tracheobronchitis (J42)",
-                "emphysema (J43.-)",
-                "lung diseases due to external agents (J60-J70)",
-            ],
-        },
-        comment="WHO browser omits some exclusions, but they're in the claml which matches the NHS browser, so we'll use the WHO value as-is",
-    ),
-    "O97": RubricChange(
-        who_2016={
-            "note": [
-                "This category is to be used to indicate death from any obstetric cause (conditions in categories O00-O75O85-O92, andO98-O99 occurring more than 42 days but less than one year after delivery. The ‘sequelae’ include conditions specified as such or as late effects, or those present one year or more after delivery."
-            ],
-            "coding-hint": [
-                "Use additional code, if desired, to identify the obstetric cause (direct or indirect)"
-            ],
-            "exclusion": [
-                "conditions specified as sequelae or late effects of obstetric causes that are not resulting in death (O94)"
-            ],
-        },
-        replace={
-            "note": {
-                "from any obstetric cause": "from sequelae any obstetric cause",
-                "(conditions in categories O00-O75O85-O92, andO98-O99": "(conditions in categories O00-O75, O85-O92, and O98-O99)",
-                "occurring more than 42 days but less than one year after delivery": "occurring more than one year after delivery",
-            },
-        },
-        comment="The WHO version is missing a closing parenthesis and also, incorrectly, says 'more than 42 days but less than one year' instead of 'more than one year'. This is a significant difference in meaning. But O96 is for 42days to 1 year, and O97 is for sequelae after 1 year. So the WHO text is probably wrong and the NHS text is correct.",
-    ),
-    # U codes
-    "U072": RubricChange(
-        who_2016={},
-        add={
-            "coding-hint": [
-                "Use this code when COVID-19 is diagnosed clinically or epidemiologically but laboratory testing is inconclusive or not available. Use additional code, if desired, to identify pneumonia or other manifestations"
-            ],
-            "exclusion": [
-                "Coronavirus infection, unspecified site (B34.2)",
-                "COVID-19: confirmed by laboratory testing (U07.1)",
-                "COVID-19: special screening examination (Z11.5)",
-                "COVID-19: suspected but ruled out by negative laboratory results (Z03.8)",
-            ],
-        },
-        comment="U072 has no rubrics in WHO so we use the NHS rubrics as-is",
-    ),
-    "U073": RubricChange(
-        who_2016={},
-        add={
-            "note": [
-                "This optional code is used to record an earlier episode of COVID-19, confirmed or probable, that influences the person’s health status, and the person no longer suffers from COVID-19. This code should not be used for primary mortality tabulation"
-            ]
-        },
-        comment="U073 has no rubrics in WHO so we use the NHS rubrics as-is",
-    ),
-    "U075": RubricChange(
-        who_2016={},
-        add={
-            "inclusion": [
-                "Cytokine storm: Temporally associated with COVID-19",
-                "Kawasaki-like syndrome: Temporally associated with COVID-19",
-                "Paediatric Inflammatory Multisystem Syndrome (PIMS): Temporally associated with COVID-19 ",
-                "Multisystem Inflammatory Syndrome in Children (MIS-C): Temporally associated with COVID-19",
-            ],
-            "exclusion": ["Mucocutaneous lymph node syndrome [Kawasaki] ( M30.3)"],
-        },
-        comment="U075 has no rubrics in WHO so we use the NHS rubrics as-is",
-    ),
-    "U076": RubricChange(
-        who_2016={},
-        add={
-            "note": [
-                "This code should not be used for international comparison or for primary mortality coding. This optional code is intended to be used when a person who may or may not be sick encounters health services for the specific purpose of receiving a COVID-19 vaccine."
-            ],
-            "coding-hint": ["Prophylactic COVID-19 vaccination"],
-            "exclusion": ["immunization not carried out (Z28.-)"],
-        },
-        comment="U076 has no rubrics in WHO so we use the NHS rubrics as-is",
-    ),
-    "U077": RubricChange(
-        who_2016={},
-        add={
-            "note": [
-                "This code is to be used as an external cause code (i.e. as a subcategory under Y59 Other and unspecified vaccines and biological substances). In addition to this, a code from another chapter of the classification should be used indicating the nature of the adverse effect"
-            ],
-            "coding-hint": [
-                "Correct administration of COVID-19 vaccine in prophylactic therapeutic use as the cause of any adverse effect."
-            ],
-        },
-        comment="U077 has no rubrics in WHO so we use the NHS rubrics as-is",
-    ),
-}
